requirements.txt
Flask>=3.1.0
gunicorn>=23.0.0
pandas>=2.2.0
gspread>=6.0.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.149.0
requests>=2.32.0
